refactor utils folder abit
